refactor(KERN): remove legacy design radio
