+++ app/components/formElements/schemaToForm/renderZodEnum.tsx
@@ -2,7 +2,7 @@ import get from "lodash/get";
 import type { z } from "zod";
 import type { StrapiFormComponent } from "~/services/cms/models/formElements/StrapiFormComponent";
 import { sortSchemaOptionsByFormComponents } from "./sortSchemaOptionsByFormComponents";
-import KernRadioGroup from "~/components/kern/formElements/KernRadioGroup";
+import KernRadioGroup from "~/components/formElements/KernRadioGroup";
 import KernTile from "~/components/kern/formElements/tile/KernTile";
 import KernCheckbox from "~/components/kern/formElements/KernCheckbox";
 import KernSelect from "~/components/kern/formElements/KernSelect";
